+++ cpp/p3/CMakeLists.txt
@@ -1,13 +1,13 @@
 file(GLOB SOURCES source/p3/*.cpp source/p3/backend/*.cpp source/p3/platform/*.cpp)
-set(SOURCES ${SOURCES})
 add_library(p3 STATIC ${SOURCES})
 target_include_directories(p3 PUBLIC source/)
 target_link_libraries(p3 
-    PUBLIC 
-    glad 
-    imgui 
-    implot 
-    glfw 
+    PUBLIC
+    lib_skia
+    glad
+    imgui
+    implot
+    glfw
     rang
     pugixml
     p3_parser
