+++ cpp/p3/source/p3/backend/OpenGL3RenderBackend.cpp
@@ -1,7 +1,7 @@
 #include <imgui.h>
 #include <backends/imgui_impl_OpenGL3.h>
 #include <glad/gl.h>
-
+#include <p3/log.h>
 #include "OpenGL3RenderBackend.h"
 #include "OpenGLRenderTarget.h"
 #include "OpenGLTexture.h"
@@ -12,6 +12,8 @@ namespace p3
     void OpenGL3RenderBackend::init()
     {
         ImGui_ImplOpenGL3_Init();
+        log_debug("creating skia context");
+        _skia_context = GrContext::MakeGL();
     }
 
     void OpenGL3RenderBackend::new_frame()
@@ -29,10 +31,10 @@ namespace p3
         _textures.push_back(std::make_unique<OpenGLTexture>());
         return _textures.back().get();
     }
-
+    
     RenderBackend::RenderTarget *OpenGL3RenderBackend::create_render_target(std::uint32_t width, std::uint32_t height)
     {
-        _render_targets.push_back(std::make_unique<OpenGLRenderTarget>(width, height));
+        _render_targets.push_back(std::make_unique<OpenGLRenderTarget>(*this, width, height));
         return _render_targets.back().get();
     }
