+++ cpp/p3/source/p3/backend/OpenGLRenderTarget.cpp
@@ -3,11 +3,12 @@
 #include <p3/log.h>
 
 #include "OpenGLRenderTarget.h"
+#include <include/gpu/gl/GrGLTypes.h>
 
 namespace p3
 {
 
-    OpenGLRenderTarget::OpenGLRenderTarget(std::uint32_t width, std::uint32_t height)
+    OpenGLRenderTarget::OpenGLRenderTarget(RenderBackend& backend, std::uint32_t width, std::uint32_t height)
         : _width(width)
         , _height(height)
     {
@@ -41,13 +42,34 @@ namespace p3
         
         log_debug("render target created ({}x{})", _width, _height);
         glPopAttrib();
+        _make_skia_surface(backend);
+    }
+
+    void OpenGLRenderTarget::_make_skia_surface(RenderBackend& backend)
+    {
+        auto framebuffer_info = GrGLFramebufferInfo{_framebuffer_id, GL_RGBA8 };
+        auto backend_render_target = GrBackendRenderTarget(_width, _height, 0, 8, framebuffer_info);
+        auto color_space = SkColorSpace::MakeSRGB();
+        _skia_surface = SkSurface::MakeFromBackendRenderTarget(
+            backend.skia_context().get(),
+            backend_render_target,
+            GrSurfaceOrigin::kTopLeft_GrSurfaceOrigin,
+            SkColorType::kRGBA_8888_SkColorType,
+            color_space,
+            nullptr);
     }
 
     OpenGLRenderTarget::~OpenGLRenderTarget()
     {
+        _skia_surface.reset();
         glDeleteFramebuffers(1, &_framebuffer_id);
         glDeleteTextures(1, &reinterpret_cast<GLuint&>(_texture_id));
         glDeleteRenderbuffers(1, &_stencil_id);
+    }
+
+    sk_sp<SkSurface> const& OpenGLRenderTarget::skia_surface() const
+    {
+        return _skia_surface;
     }
 
     unsigned int OpenGLRenderTarget::framebuffer_id() const
