+++ cpp/p3_python/source/p3/python/Menu.cpp
@@ -1,4 +1,3 @@
-
 #include "p3ui.h"
 #include <p3/Menu.h>
